requirements.txt
# Together AI SDK (latest version)
together>=1.3.0

# Async support
aiohttp>=3.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0

# Web dashboard
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Token counting (for accurate token generation)
tiktoken>=0.5.0

# Additional utilities
aiofiles>=23.0.0

# Environment variable to hide Together SDK banner
# Add to your shell: export TOGETHER_NO_BANNER=1
